+++ Numeric/GaussQuadratureLin.cpp
@@ -13,7 +13,8 @@ IntPt *getGQLPts(int order)
 {
   // Number of Gauss Point: (order + 1) / 2 *ROUNDED UP*
   int n = (order + 1) / (double)2 + 0.5;
-  if(GQL.size() < order + 1) GQL.resize(order + 1, nullptr);
+  if(static_cast<int>(GQL.size()) < order + 1)
+    GQL.resize(order + 1, nullptr);
   if(!GQL[order]) {
     double *pt, *wt;
     gmshGaussLegendre1D(n, &pt, &wt);
